requirements.txt
yt-dlp==2023.11.16
tkinter-page==0.5.1
python-dotenv==1.0.0
customtkinter==5.2.0
Pillow==10.0.0
ffmpeg-python==0.2.0
